Fixing string decimal conversion

+++ ItemInfo.cs
@@ -10,7 +10,7 @@ namespace DynamicWebScrape
     {
         public string ItemName { get; set; }
         public string ItemURL { get; set; }
-        public decimal ItemPrice { get; set; }
+        public double ItemPrice { get; set; }
         public int InternetNum { get; set; }
         public int Id { get; set; }
     }
